+++ hibernate-rx-core/src/test/java/org/hibernate/rx/MySQLAutoincrementTest.java
@@ -185,7 +185,7 @@ public class MySQLAutoincrementTest extends BaseRxTest {
 	public static class Basic {
 
 		@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-		Integer id;
+		Long id;
 		@Version Integer version;
 		String string;
 
@@ -240,7 +240,7 @@ public class MySQLAutoincrementTest extends BaseRxTest {
 			this.string = string;
 		}
 
-		public Basic(Integer id, String string) {
+		public Basic(Long id, String string) {
 			this.id = id;
 			this.string = string;
 		}
@@ -282,11 +282,11 @@ public class MySQLAutoincrementTest extends BaseRxTest {
 			loaded = true;
 		}
 
-		public Integer getId() {
+		public Long getId() {
 			return id;
 		}
 
-		public void setId(Integer id) {
+		public void setId(Long id) {
 			this.id = id;
 		}
